refactor(struct): extract EPUB profile asset tests
